# server/controllers/IQ_db/generated/PlayersControllerGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE PlayersControllerGenerated.js PLEASE EDIT ../PlayersController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import PlayersModel from "../../../models/IQ_db/PlayersModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import PlayersController from "../PlayersController";

const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/players`;
    router.post(baseUrl + "", authorize([]), PlayersController.create);
    router.delete(baseUrl + "/:id", authorize([]), PlayersController.delete);
    router.get(baseUrl + "/findByName/:key", authorize([]), PlayersController.findByName);
    router.get(baseUrl + "/:id", authorize([]), PlayersController.get);
    router.get(baseUrl + "", authorize([]), PlayersController.list);
    router.post(baseUrl + "/:id", authorize([]), PlayersController.update);
  },


  // CRUD METHODS


  /**
  * PlayersModel.create
  *   @description CRUD ACTION create
  *
  */
  create: async (req, res) => {
    try {
      const result = await PlayersModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * PlayersModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id
  *
  */
  delete: async (req, res) => {
    try {
      const result = await PlayersModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * PlayersModel.findByName
  *   @description CRUD ACTION findByName
  *   @param String key Id of the resource Name to search
  *
  */
  findByName: async (req, res) => {
    try {
      const result = await PlayersModel.findByName(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * PlayersModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id Players
  *   @returns Players
  *
  */
  get: async (req, res) => {
    try {
      const result = await PlayersModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * PlayersModel.list
  *   @description CRUD ACTION list
  *   @returns ARRAY OF Players
  *
  */
  list: async (req, res) => {
    try {
      const result = await PlayersModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * PlayersModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id Players
  *   @returns Players
  *
  */
  update: async (req, res) => {
    try {
      const result = await PlayersModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;