# server/swagger.yaml
openapi: 3.0.0
info:
  title: IQ
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /matches/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Matches
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Matches
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matches"
    post:
      summary: CRUD ACTION update
      tags:
        - Matches
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matches"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Matches"
    
  #-----  Players  -----
...
  /players/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Players
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Players
      parameters:
        - name: id
          in: path
          description: Id Players
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Players"
    post:
      summary: CRUD ACTION update
      tags:
        - Players
      parameters:
        - name: id
          in: path
          description: Id Players
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Players"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Players"
  /players/findByName/{key}:
    get:
      summary: CRUD ACTION findByName
      tags:
        - Players
      parameters:
        - name: key
          in: path
          description: Id of the resource Name to search
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Players"
    
  #-----  Preferences  -----
...
  /teams/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Teams
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Teams
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teams"
    post:
      summary: CRUD ACTION update
      tags:
        - Teams
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teams"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Teams"
  /teams/findByname/{key}:
    get:
      summary: CRUD ACTION findByname
      tags:
        - Teams
      parameters:
        - name: key
          in: path
          description: Id of the resource name to search
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teams"
    
  #-----  User  -----
...
  /user/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
    
  #-----  User  -----
  /Users/{id}/changePassword:
    post:
      summary: Change password by admin API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordAdmin:
                  type: string
      responses:
        '200':
          description: OK
  /Users/{id}:
    post:
      summary: Update User API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    delete:
      summary: Delete User API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Matches:
      properties:
        _id:
          type: string
        away_team:
          type: string
        home_team:
          type: string
        league_id:
          type: integer
        match_id:
          type: integer
        match_start:
          type: string
        status:
          type: integer
        status_code:
          type: integer
      required:
        -  _id
    Players:
      properties:
        _id:
          type: string
        Age:
          type: number
        Equip:
          type: string
        Height:
          type: number
        Name:
          type: string
        Number:
          type: number
        Position:
          type: string
        Weight:
          type: number
      required:
        -  _id
        - Age
        - Equip
        - Height
        - Name
        - Number
        - Position
        - Weight
...
    Teams:
      properties:
        _id:
          type: string
        color1:
          type: string
        color2:
          type: string
        country:
          type: string
        founded:
          type: string
        logo:
          type: string
        name:
          type: string
      required:
        -  _id
        - color1
        - color2
        - country
        - logo
        - name
    User:
      properties:
        _id:
          type: string
        email:
          type: string
        name:
          type: string
        password:
          type: string
        roles:
          type: string
        surname:
          type: string
        username:
          type: string
      required:
        -  _id
        - password
        - username